moonlight/requirements.txt
# MoonLight Core Dependencies
# Python 3.10+ required

# Core Framework
asyncio-mqtt==0.16.1
aiofiles==23.2.1
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1

# HTTP & WebSocket
aiohttp==3.9.1
websockets==12.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data & Analytics
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
ta-lib==0.4.28  # Teknik analiz (opsiyonel, binary gerektirir)

# Machine Learning (opsiyonel)
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Security & Secrets
cryptography==41.0.7
keyring==24.3.0
pywin32==306  # Windows için

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Type Checking
mypy==1.7.1
types-PyYAML==6.0.12.12

# Code Quality
ruff==0.1.6
black==23.11.0

# Database
sqlalchemy==2.0.23  # Opsiyonel ORM
alembic==1.13.0     # Migrasyon

# Forex/Crypto Data (opsiyonel)
ccxt==4.1.58        # Multi-exchange (opsiyonel)

# Performance
uvloop==0.19.0      # Linux için (Windows'ta opsiyonel)
orjson==3.9.10      # Hızlı JSON

# Prometheus (opsiyonel)
prometheus-client==0.19.0
